pythongit/imported_recruitflow/RecruitFlow/RecruitFlow/python-services/requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
pdfplumber==0.11.8
python-docx==1.2.0
pytest==9.0.1
httpx==0.28.1
python-multipart==0.0.20
PyPDF2==3.0.1